+++ NEWS.md
@@ -5,6 +5,7 @@
 - Added dtm_align
 - Added dtm_sample
 - cbind_morphological gains argument which to specify which morphological features to extract
+- txt_count now returns NA when NA is provided instead of an error
 
 ## CHANGES IN udpipe VERSION 0.8.4-1