+++ NEWS.md
@@ -4,6 +4,7 @@
 - Added groups argument to dtm_colsums and dtm_rowsums
 - Added dtm_align
 - Added dtm_sample
+- Added document_term_matrix.matrix
 - cbind_morphological gains argument which to specify which morphological features to extract
 - txt_count now returns NA when NA is provided instead of an error
 - txt_contains now returns NA when NA is provided instead of FALSE, unless value is set to TRUE