+++ NEWS.md
@@ -8,6 +8,7 @@
 - cbind_morphological gains argument which to specify which morphological features to extract
 - txt_count now returns NA when NA is provided instead of an error
 - txt_contains now returns NA when NA is provided instead of FALSE, unless value is set to TRUE
+- paste.data.frame now works as well if a data.table is passed instead of a data.frame
 
 ## CHANGES IN udpipe VERSION 0.8.4-1